add space parallelisation
